Move the endpoint tables and the transport layer into probe-rs-rpc
