+++ probe-rs-rpc/src/transport/mod.rs
@@ -1,7 +1,5 @@
-//! Postcard RPC wire transport implementation, using `futures_util::Sink` and
-//! `futures_util::Stream` types as the underlying transport, and 4-byte length prefix encoding.
-
 pub mod memory;
+
 #[cfg(all(feature = "remote", unix))]
 pub mod unix;
 #[cfg(feature = "remote")]
