+++ probe-rs-rpc/src/transport/websocket.rs
@@ -1,4 +1,3 @@
-#[cfg(feature = "remote")]
 use axum::extract::ws;
 use futures_util::{FutureExt, Sink, SinkExt, Stream, StreamExt};
 use postcard_rpc::server::{WireRxErrorKind, WireTxErrorKind};
@@ -12,7 +11,7 @@ use tokio::sync::Mutex;
 use tokio_tungstenite::tungstenite::Message;
 use tokio_util::bytes::Bytes;
 
-use crate::rpc::transport::{
+use crate::transport::{
     Deframer, frame,
     memory::{PostcardReceiver, PostcardSender},
 };
@@ -83,7 +82,7 @@ pub struct WebsocketTx<S> {
     writer: Arc<Mutex<S>>,
 }
 impl<S> WebsocketTx<S> {
-    pub(crate) fn new(writer: S) -> Self {
+    pub fn new(writer: S) -> Self {
         Self {
             writer: Arc::new(Mutex::new(writer)),
         }
@@ -113,12 +112,11 @@ pub struct AxumWebsocketTx<S> {
     writer: S,
 }
 impl<S> AxumWebsocketTx<S> {
-    pub(crate) fn new(writer: S) -> Self {
+    pub fn new(writer: S) -> Self {
         Self { writer }
     }
 }
 
-#[cfg(feature = "remote")]
 impl<S> Sink<Vec<u8>> for AxumWebsocketTx<S>
 where
     S: Sink<ws::Message> + Unpin,
